feat: add scanner support to vulnscan manager
This commit adds the ability to specify a scanner when creating a
vulnerability scan. It also adds a new method to the VulnscanManager
class, get_scanners(), which returns a list of available scanners.

The default port list is changed to "all iana assigned tcp" to ensure
that all relevant ports are scanned by default.